+++ axum/src/db/crud/mod.rs
@@ -1,3 +1,3 @@
 pub mod file;
+pub mod learning;
 pub mod task;
-pub use file::*;
